use Flow Grid APIs switch StorefrontI view APIs (#259)
* update Flow and vaadin-grid to enable Java Grid API for Storefront

Reasons to update:
 - switching the Storefront view onto the Grid Java API requires Component Renderer support
 - Flow version 1.0.0.alpha8 or newer is required to use the Grid Component Renderer
 - because of the bugs in alpha8, the actually required Flow version is 1.0.0.alpha9
 - Flow version 1.0.0.alpha9 requires vaadin-grid 4.1.0

Until a version of vaadin-bom including Flow 1.0.0-alpha9 is released, switching to the SNAPSHOT.

Upgrade of vaadin-grid to 4.1.0 is done by updating the version of the 'vaadin' meta-dependency to the latest (10.0.0-alpha5; that includes vaadin-grid 4.1.0-alpha1) and then still bumping the version of vaadin-grid to the latest available: 4.1.0-alpha3.

After the 'vaadin' meta-dependency is updated, there is no need to explicitly set the versions for other vaadin elements (except the grid) because the correct versions are already included into the meta dependency.

* switch to Java API to create the grid on the Storefront view

Done:
 - server-side Java instance of Grid is used to create and configure the grid
 - the StorefrontItemDetailWrapper component is used to render grid items (created and configured on the server-side)
 - visually the grid still looks the same

TODO:
 - switch to Java for event handling
 - switch to Java for data binding
 - fix headers (for the StorefrontItemDetailWrapper items)
 - use the grid data provider API

* convert Storefront headers bindings from JS to Java

fixes BFF-368: removes custom JSON serialization

* remove conditional Polymer template (to work around a Flow limitation)

Flow has a very limited support for conditional Polymer templates. In this case there was a runtime exception:

`java.lang.IllegalStateException: Couldn't parse the template: sub-templates are not supported. Sub-template found: <order-detail item="[[order]]" desktop-view="[[desktopView]]"></order-detail>`

To avoid it, switch from a conditional template onto using the 'hidden' HTML attribute.

* fix the production build and tests

The Grid Component Renderer uses an internal Polymer component defined in a file inside flow-server.jar. In production builds (with bundling) this file should be included into the bundle (and therefore should be availabe in the webapp folder for `polymer build`).

* exclude flow-component-renderer.html from linting

The flow-component-renderer.html file is not maintained by the Bakery project and therefore we do not care whether or not it confirms to the linter rules. If it does not, that should not be a build failure.

* switch the StorefrontItemDetailWrapper component to Java (from JS)

Create Java APIs for the child components of StorefrontItemDetailWrapper:
 - update the Java API for the OrderDetails component (make the Java CommentEvent functionally equivalent to its JS counterpart)
 - create a Java API for the StorefrontItem component (with the only method: `setOrder(order)`)

In StorefrontItemDetailWrapper use StorefrontItem and OrderDetail via their Java APIs:
 - get references to the Java components with the `@Id` annotation
 - remove the 'order' property from the StorefrontItemDetailWrapper model

In StorefrontView update the presenter methods handling the StorefrontItemDetailWrapper component.

TODO:
- update the Storefront view to use the Java API with the OrderDetail component also in the mobile view
- remove the JS API from the OrderDetails component
   - remove the JS 'comment-added' event
   - remove the JS 'edit' and 'close' events

* switch the Storefront view to Java APIs for the mobile view

Key changes:
 - use the same OrderDetail component in both the mobile and the desktop modes of the Storefront view
 - use the Java APIs of the OrderDetail component also when in the mobile mode
 - remove the JS APIs of the order-detail element (now unused)
 - remove the order-detail-mobile-dialog element (became unused)

update the touch scrolling on/off on logic on the Storefront view
 - remove the no-longer-used `_selectedItem` and `_editableItem` properties
 - enable touch scrolling whenever the editor is not opened

* fix the double triggering of the Closed event for the StorefrontItemDetailWrapper

Separate the 'opened' and 'closed' events from the item-detail-inline element from the StorefrontItemDetailWrapper events.
Create the events specific for the storefront-item-detail-wrapper element: 'expanded' and 'collapsed'.

* downgrade flow-spring-addon to 1.0.0.alpha7 to avoid app termination deadlock

newer versions of the addon have this issue: https://github.com/vaadin/flow/issues/2871

* use vaadin platform 10.0.0-alpha6 instead of the snapshot

the latest release of the vaadin platform BOM includes Flow 1.0.0.alpha9 which was the reason to use SNAPSHOT