create a common root layout for all views of the app
The BakeryApp class is split into two: BakeryApp and BakeryLayout.
All views that had `BakeryApp` as the parent layout now have `BakeryLayout` as a parent instead.
All views that did not have a parent layout (i.e. the LoginView) now have `BakeryApp` as a parent.

This allows having a single place where to place the initial page configuration so that it applies to all views.

Jira: BFF-455